fix: move git deps to npm
