+++ __mocks__/@synonymdev/react-native-keychain/index.ts
