+++ package.json
@@ -50,6 +50,8 @@
     "@synonymdev/blocktank-lsp-http-client": "1.5.0",
     "@synonymdev/feeds": "3.0.0",
     "@synonymdev/ledger": "0.0.5",
+    "@synonymdev/react-native-change-icon": "5.0.0",
+    "@synonymdev/react-native-keychain": "8.2.2",
     "@synonymdev/react-native-ldk": "0.0.150",
     "@synonymdev/react-native-lnurl": "0.0.10",
     "@synonymdev/result": "0.0.2",
@@ -87,7 +89,6 @@
     "react-native-address-generator": "0.3.3",
     "react-native-biometrics": "3.0.1",
     "react-native-camera-kit": "14.0.0-beta15",
-    "react-native-change-icon": "pwltr/react-native-change-icon",
     "react-native-device-info": "11.1.0",
     "react-native-dotenv": "3.4.11",
     "react-native-draggable-flatlist": "4.0.1",
@@ -98,7 +99,6 @@
     "react-native-haptic-feedback": "2.2.0",
     "react-native-image-picker": "7.1.2",
     "react-native-keyboard-accessory": "0.1.16",
-    "react-native-keychain": "synonymdev/react-native-keychain#refactor/android-data-store",
     "react-native-localize": "3.0.2",
     "react-native-mmkv": "2.12.2",
     "react-native-modal": "13.0.1",
