+++ src/utils/lightning/index.ts
@@ -1,5 +1,5 @@
 import { EmitterSubscription, InteractionManager } from 'react-native';
-import Keychain from 'react-native-keychain';
+import Keychain from '@synonymdev/react-native-keychain';
 import * as bitcoin from 'bitcoinjs-lib';
 import RNFS from 'react-native-fs';
 import { err, ok, Result } from '@synonymdev/result';
